requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
requests==2.32.3
feedparser==6.0.11
openai==1.12.0
